requirements.txt
boto3==1.9.106
botocore==1.12.107
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.1
djangorestframework==3.9.2
docutils==0.14
geojson==2.4.1
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
numpy==1.16.2
Pillow==5.2.0
psycopg2==2.7.7
pyowm==2.10.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
timezonefinderL==2.0.1
urllib3==1.24.1
whitenoise==4.1.2
